refactor: rename daemon→loop; surface as `healthd run`
The `daemon` command path was a vestige from when healthd managed its
own LaunchAgent lifecycle. That code was removed in 63a396a — supervision
is now external (process-compose / systemd / launchd). The remaining
"daemon run" subcommand was just running a foreground loop, so the
naming became misleading.

- internal/daemon/ → internal/loop/; package daemon → package loop;
  RunLoop → Run; matching test name updates.
- cmd/daemon.go → cmd/run.go; removed the parent `daemon` command;
  exposed `healthd run` at the top level.
- ARCHITECTURE.md mermaid diagram and component table updated to use
  the loop package and `healthd run` invocation.
- README, AGENTS.md, operator skill, and openclaw integration doc all
  updated to reference the new command path.

Breaking change for anyone invoking `healthd daemon run` directly —
update supervisor configs to `healthd run`.
