+++ internal/loop/coverage_test.go
@@ -1,4 +1,4 @@
-package daemon
+package loop
 
 import (
 	"bytes"
@@ -13,11 +13,11 @@ import (
 	"github.com/uinaf/healthd/internal/config"
 )
 
-func TestRunLoopAdditionalBranches(t *testing.T) {
+func TestRunAdditionalBranches(t *testing.T) {
 	homeDir := t.TempDir()
 	t.Setenv("HOME", homeDir)
 
-	if err := RunLoop(context.Background(), config.Config{Interval: "bad"}, io.Discard); err == nil || !strings.Contains(err.Error(), "parse schedule interval") {
+	if err := Run(context.Background(), config.Config{Interval: "bad"}, io.Discard); err == nil || !strings.Contains(err.Error(), "parse schedule interval") {
 		t.Fatalf("expected interval parse error, got %v", err)
 	}
 
@@ -27,7 +27,7 @@ func TestRunLoopAdditionalBranches(t *testing.T) {
 		Checks:   []config.CheckConfig{{Name: "ok", Command: "true"}},
 		Notify:   config.NotifyConfig{Cooldown: "bad"},
 	}
-	if err := RunLoop(context.Background(), cfgCooldown, io.Discard); err == nil || !strings.Contains(err.Error(), "parse cooldown") {
+	if err := Run(context.Background(), cfgCooldown, io.Discard); err == nil || !strings.Contains(err.Error(), "parse cooldown") {
 		t.Fatalf("expected cooldown parse error, got %v", err)
 	}
 
@@ -39,7 +39,7 @@ func TestRunLoopAdditionalBranches(t *testing.T) {
 			Type: "unsupported",
 		}}},
 	}
-	if err := RunLoop(context.Background(), cfgBackend, io.Discard); err == nil || !strings.Contains(err.Error(), "unsupported backend type") {
+	if err := Run(context.Background(), cfgBackend, io.Discard); err == nil || !strings.Contains(err.Error(), "unsupported backend type") {
 		t.Fatalf("expected unsupported backend error, got %v", err)
 	}
 
@@ -56,8 +56,8 @@ func TestRunLoopAdditionalBranches(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 45*time.Millisecond)
 	defer cancel()
 	var out bytes.Buffer
-	if err := RunLoop(ctx, cfgRun, &out); err != nil {
-		t.Fatalf("RunLoop() error = %v", err)
+	if err := Run(ctx, cfgRun, &out); err != nil {
+		t.Fatalf("Run() error = %v", err)
 	}
 	if !strings.Contains(out.String(), "notify dispatch error for failing") {
 		t.Fatalf("expected dispatch error output, got %q", out.String())
